tenant and table test

+++ packages/saltcorn-data/tests/tenant.test.js
@@ -6,16 +6,22 @@ const { createTenant, deleteTenant } = require("../models/tenant");
 afterAll(db.close);
 
 beforeAll(async () => {
-  await db.query(`drop schema if exists test10 CASCADE `);
+  if (!db.isSQLite) await db.query(`drop schema if exists test10 CASCADE `);
 });
 
 describe("Tenant", () => {
-  it("can create a new tenant", async () => {
-    db.enable_multi_tenant();
-    await createTenant("test10");
-  });
-  it("can delete a tenant", async () => {
-    db.enable_multi_tenant();
-    await deleteTenant("test10");
-  });
+  if (!db.isSQLite) {
+    it("can create a new tenant", async () => {
+      db.enable_multi_tenant();
+      await createTenant("test10");
+    });
+    it("can delete a tenant", async () => {
+      db.enable_multi_tenant();
+      await deleteTenant("test10");
+    });
+  } else {
+    it("does not support tenants on SQLite", async () => {
+      expect(db.isSQLite).toBe(true);
+    });
+  }
 });
